Victory condition tests and more files
Made a victory condition test (even though I decided to not change that
function).  Moved it to a separate cpp/h file.  This required me to move
my class to a separate cpp/h file so the test files could know the
class.  Seems inelegant to do it this way since only TicTacToe will use
most of this but hopefully the more I do the more I recognize where it
really helps.